refactor: application 모듈 영역 적용
