+++ kitchenpos-application/src/main/java/kitchenpos/ui/order/OrderRestController.java
@@ -1,0 +1,38 @@
+package kitchenpos.ui.order;
+
+import kitchenpos.order.application.OrderService;
+import kitchenpos.order.domain.OrderStatus;
+import kitchenpos.order.dto.OrderRequest;
+import kitchenpos.order.dto.OrderResponse;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.net.URI;
+import java.util.List;
+
+@RestController
+@RequestMapping("/api/orders")
+public class OrderRestController {
+    private final OrderService orderService;
+
+    public OrderRestController(OrderService orderService) {
+        this.orderService = orderService;
+    }
+
+    @PostMapping
+    public ResponseEntity<OrderResponse> create(@RequestBody OrderRequest request) {
+        OrderResponse created = orderService.create(request);
+        return ResponseEntity.created(URI.create("/api/orders/" + created.getId())).body(created);
+    }
+
+    @GetMapping
+    public ResponseEntity<List<OrderResponse>> list() {
+        return ResponseEntity.ok().body(orderService.list());
+    }
+
+    @PutMapping("/{orderId}/order-status")
+    public ResponseEntity<OrderResponse> changeOrderStatus(@PathVariable Long orderId,
+                                                           @RequestParam OrderStatus request) {
+        return ResponseEntity.ok(orderService.changeOrderStatus(orderId, request));
+    }
+}
