+++ kitchenpos-application/src/main/java/kitchenpos/ui/tablegroup/TableGroupRestController.java
@@ -1,0 +1,31 @@
+package kitchenpos.ui.tablegroup;
+
+import kitchenpos.tablegroup.application.TableGroupService;
+import kitchenpos.tablegroup.dto.TableGroupRequest;
+import kitchenpos.tablegroup.dto.TableGroupResponse;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.net.URI;
+
+@RestController
+@RequestMapping("/api/table-groups")
+public class TableGroupRestController {
+    private final TableGroupService tableGroupService;
+
+    public TableGroupRestController(TableGroupService tableGroupService) {
+        this.tableGroupService = tableGroupService;
+    }
+
+    @PostMapping
+    public ResponseEntity<TableGroupResponse> create(@RequestBody TableGroupRequest request) {
+        TableGroupResponse created = tableGroupService.create(request);
+        return ResponseEntity.created(URI.create("/api/table-groups/" + created.getId())).body(created);
+    }
+
+    @DeleteMapping("/{tableGroupId}")
+    public ResponseEntity<Void> ungroup(@PathVariable Long tableGroupId) {
+        tableGroupService.ungroup(tableGroupId);
+        return ResponseEntity.noContent().build();
+    }
+}
