--- src/main/java/kitchenpos/menu/ui/MenuGroupRestController.java
@@ -1,35 +1,0 @@
-package kitchenpos.menu.ui;
-
-import kitchenpos.menu.application.MenuGroupService;
-import kitchenpos.menu.dto.MenuGroupRequest;
-import kitchenpos.menu.dto.MenuGroupResponse;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.net.URI;
-import java.util.List;
-
-@RestController
-@RequestMapping("/api/menu-groups")
-public class MenuGroupRestController {
-    private final MenuGroupService menuGroupService;
-
-    public MenuGroupRestController(MenuGroupService menuGroupService) {
-        this.menuGroupService = menuGroupService;
-    }
-
-    @PostMapping
-    public ResponseEntity<MenuGroupResponse> create(@RequestBody MenuGroupRequest request) {
-        MenuGroupResponse created = menuGroupService.create(request);
-        return ResponseEntity.created(URI.create("/api/menu-groups/" + created.getId())).body(created);
-    }
-
-    @GetMapping
-    public ResponseEntity<List<MenuGroupResponse>> list() {
-        return ResponseEntity.ok().body(menuGroupService.list());
-    }
-}
