--- src/main/java/kitchenpos/menu/ui/ProductRestController.java
@@ -1,36 +1,0 @@
-package kitchenpos.menu.ui;
-
-import kitchenpos.menu.application.ProductService;
-import kitchenpos.menu.dto.ProductRequest;
-import kitchenpos.menu.dto.ProductResponse;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.net.URI;
-import java.util.List;
-
-@RestController
-@RequestMapping("/api/products")
-public class ProductRestController {
-    private final ProductService productService;
-
-    public ProductRestController(ProductService productService) {
-        this.productService = productService;
-    }
-
-    @PostMapping
-    public ResponseEntity<ProductResponse> create(@RequestBody ProductRequest request) {
-        ProductResponse created = productService.create(request);
-        return ResponseEntity.created(URI.create("/api/products/" + created.getId())).body(created);
-    }
-
-    @GetMapping
-    public ResponseEntity<List<ProductResponse>> list() {
-        return ResponseEntity.ok().body(productService.list());
-    }
-}
-
