--- src/main/java/kitchenpos/tablegroup/ui/TableRestController.java
@@ -1,45 +1,0 @@
-package kitchenpos.tablegroup.ui;
-
-import kitchenpos.tablegroup.application.TableService;
-import kitchenpos.tablegroup.domain.OrderTableEmpty;
-import kitchenpos.tablegroup.domain.OrderTableGuests;
-import kitchenpos.tablegroup.dto.OrderTableRequest;
-import kitchenpos.tablegroup.dto.OrderTableResponse;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.net.URI;
-import java.util.List;
-
-@RestController
-@RequestMapping("/api/tables")
-public class TableRestController {
-    private final TableService tableService;
-
-    public TableRestController(TableService tableService) {
-        this.tableService = tableService;
-    }
-
-    @PostMapping
-    public ResponseEntity<OrderTableResponse> create(@RequestBody OrderTableRequest request) {
-        OrderTableResponse created = tableService.create(request);
-        return ResponseEntity.created(URI.create("/api/tables/" + created.getId())).body(created);
-    }
-
-    @GetMapping
-    public ResponseEntity<List<OrderTableResponse>> list() {
-        return ResponseEntity.ok().body(tableService.list());
-    }
-
-    @PutMapping("/{orderTableId}/empty")
-    public ResponseEntity<OrderTableResponse> changeEmpty(@PathVariable Long orderTableId,
-                                                          @RequestBody OrderTableEmpty request) {
-        return ResponseEntity.ok().body(tableService.changeEmpty(orderTableId, request));
-    }
-
-    @PutMapping("/{orderTableId}/number-of-guests")
-    public ResponseEntity<OrderTableResponse> changeNumberOfGuests(@PathVariable Long orderTableId,
-                                                                   @RequestBody OrderTableGuests request) {
-        return ResponseEntity.ok().body(tableService.changeNumberOfGuests(orderTableId, request));
-    }
-}
